Some refactoring in storage file
